Clean up preference bindings further
