+++ data/binding/items.go
@@ -2,7 +2,6 @@ package binding
 
 import (
 	"bytes"
-	"sync/atomic"
 
 	"fyne.io/fyne/v2"
 	"fyne.io/fyne/v2/storage"
@@ -320,55 +319,4 @@ func (b *baseExternalItem[T]) Set(val T) error {
 
 func (b *baseExternalItem[T]) Reload() error {
 	return b.Set(*b.val)
-}
-
-type prefBoundBase[T bool | float64 | int | string] struct {
-	base
-	key   string
-	get   func(string) T
-	set   func(string, T)
-	cache atomic.Pointer[T]
-}
-
-func (b *prefBoundBase[T]) Get() (T, error) {
-	cache := b.get(b.key)
-	b.cache.Store(&cache)
-	return cache, nil
-}
-
-func (b *prefBoundBase[T]) Set(v T) error {
-	b.set(b.key, v)
-
-	b.lock.RLock()
-	defer b.lock.RUnlock()
-	b.trigger()
-	return nil
-}
-
-func (b *prefBoundBase[T]) setKey(key string) {
-	b.key = key
-}
-
-func (b *prefBoundBase[T]) checkForChange() {
-	val := b.cache.Load()
-	if val != nil && b.get(b.key) == *val {
-		return
-	}
-	b.trigger()
-}
-
-func lookupExistingBinding[T any](key string, p fyne.Preferences) (bindableItem[T], bool) {
-	binds := prefBinds.getBindings(p)
-	if binds == nil {
-		return nil, false
-	}
-
-	if listen, ok := binds.Load(key); listen != nil && ok {
-		if l, ok := listen.(bindableItem[T]); ok {
-			return l, ok
-		}
-		fyne.LogError(keyTypeMismatchError+key, nil)
-	}
-
-	return nil, false
 }
